Clases v2.2, Main v1.5: Main - Agregado menu de combate funcional, falta menu de pueblos y manejar niveles, terminado en un 95%. Clases - Agregadas clases item, potion, powder y elixir. Funcionalidad de los items creada, validada y probada. Hecho forward declaration para asegurar el correcto funcionamiento de estas

+++ armadura.h
@@ -7,7 +7,7 @@ class armadura{
  protected:
 	double defense;
 	string atribute;
-	int durability, currDur;
+	double durability, currDur;
  public:
 	armadura(string,double=100.0);
 	string getAtribute()const;
